+++ core/src/main/kotlin/net/corda/core/crypto/CryptoUtils.kt
@@ -2,7 +2,6 @@
 
 package net.corda.core.crypto
 
-import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.AnonymousParty
 import net.corda.core.identity.Party
 import net.corda.core.serialization.CordaSerializable