+++ core/src/main/kotlin/net/corda/flows/FinalityFlow.kt
@@ -89,7 +89,7 @@ class FinalityFlow(val transactions: Iterable<SignedTransaction>,
 
     private fun needsNotarySignature(stx: SignedTransaction): Boolean {
         val wtx = stx.tx
-        val needsNotarisation = wtx.inputs.isNotEmpty() || wtx.timestamp != null
+        val needsNotarisation = wtx.inputs.isNotEmpty() || wtx.timeWindow != null
         return needsNotarisation && hasNoNotarySignature(stx)
 
     }