+++ finance/src/main/kotlin/net/corda/flows/TwoPartyTradeFlow.kt
@@ -225,10 +225,10 @@ object TwoPartyTradeFlow {
             tx.addOutputState(state, tradeRequest.assetForSale.state.notary)
             tx.addCommand(command, tradeRequest.assetForSale.state.data.owner.owningKey)
 
-            // And add a request for timestamping: it may be that none of the contracts need this! But it can't hurt
-            // to have one.
+            // And add a request for a time-window: it may be that none of the contracts need this!
+            // But it can't hurt to have one.
             val currentTime = serviceHub.clock.instant()
-            tx.setTime(currentTime, 30.seconds)
+            tx.addTimeWindow(currentTime, 30.seconds)
             return Pair(tx, cashSigningPubKeys)
         }
         // DOCEND 1