+++ node/src/test/kotlin/net/corda/node/services/database/RequeryConfigurationTest.kt
@@ -167,7 +167,7 @@ class RequeryConfigurationTest {
                 notary = DUMMY_NOTARY,
                 signers = emptyList(),
                 type = TransactionType.General,
-                timestamp = null
+                timeWindow = null
         )
         return SignedTransaction(wtx.serialized, listOf(DigitalSignature.WithKey(NullPublicKey, ByteArray(1))))
     }