+++ samples/irs-demo/src/main/kotlin/net/corda/irs/contract/IRS.kt
@@ -6,7 +6,6 @@ import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.containsAny
 import net.corda.core.flows.FlowLogicRefFactory
 import net.corda.core.identity.AbstractParty
-import net.corda.core.identity.AnonymousParty
 import net.corda.core.identity.Party
 import net.corda.core.node.services.ServiceType
 import net.corda.core.serialization.CordaSerializable
@@ -459,7 +458,7 @@ class InterestRateSwap : Contract {
                 fixingCalendar, index, indexSource, indexTenor)
     }
 
-    override fun verify(tx: TransactionForContract) = verifyClause(tx, AllOf(Clauses.Timestamped(), Clauses.Group()), tx.commands.select<Commands>())
+    override fun verify(tx: TransactionForContract) = verifyClause(tx, AllOf(Clauses.TimeWindow(), Clauses.Group()), tx.commands.select<Commands>())
 
     interface Clauses {
         /**
@@ -515,13 +514,13 @@ class InterestRateSwap : Contract {
             }
         }
 
-        class Timestamped : Clause<ContractState, Commands, Unit>() {
+        class TimeWindow : Clause<ContractState, Commands, Unit>() {
             override fun verify(tx: TransactionForContract,
                                 inputs: List<ContractState>,
                                 outputs: List<ContractState>,
                                 commands: List<AuthenticatedObject<Commands>>,
                                 groupingKey: Unit?): Set<Commands> {
-                require(tx.timestamp?.midpoint != null) { "must be timestamped" }
+                require(tx.timeWindow?.midpoint != null) { "must be have a time-window)" }
                 // We return an empty set because we don't process any commands
                 return emptySet()
             }