+++ samples/irs-demo/src/main/kotlin/net/corda/irs/flows/UpdateBusinessDayFlow.kt
@@ -1,10 +1,10 @@
 package net.corda.irs.flows
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.core.identity.Party
 import net.corda.core.flows.FlowLogic
 import net.corda.core.flows.InitiatingFlow
 import net.corda.core.flows.StartableByRPC
+import net.corda.core.identity.Party
 import net.corda.core.node.CordaPluginRegistry
 import net.corda.core.node.NodeInfo
 import net.corda.core.node.PluginServiceHub
@@ -66,7 +66,7 @@ object UpdateBusinessDayFlow {
         /**
          * Returns recipients ordered by legal name, with notary nodes taking priority over party nodes.
          * Ordering is required so that we avoid situations where on clock update a party starts a scheduled flow, but
-         * the notary or counterparty still use the old clock, so the timestamp on the transaction does not validate.
+         * the notary or counterparty still use the old clock, so the time-window on the transaction does not validate.
          */
         private fun getRecipients(): Iterable<NodeInfo> {
             val notaryNodes = serviceHub.networkMapCache.notaryNodes