+++ samples/simm-valuation-demo/src/main/kotlin/net/corda/vega/flows/StateRevisionFlow.kt
@@ -8,7 +8,6 @@ import net.corda.core.transactions.SignedTransaction
 import net.corda.flows.AbstractStateReplacementFlow
 import net.corda.flows.StateReplacementException
 import net.corda.vega.contracts.RevisionedState
-import java.security.PublicKey
 
 /**
  * Flow that generates an update on a mutable deal state and commits the resulting transaction reaching consensus
@@ -20,7 +19,7 @@ object StateRevisionFlow {
         override fun assembleTx(): Pair<SignedTransaction, List<AbstractParty>> {
             val state = originalState.state.data
             val tx = state.generateRevision(originalState.state.notary, originalState, modification)
-            tx.setTime(serviceHub.clock.instant(), 30.seconds)
+            tx.addTimeWindow(serviceHub.clock.instant(), 30.seconds)
 
             val stx = serviceHub.signInitialTransaction(tx)
             return Pair(stx, state.participants)